test adding network code to dictionnary in JS

+++ web/test.js
@@ -242,7 +242,7 @@ function receiver(msg)
     console.log(view[0]);
     var view = new DataView(msg);
     console.log(view);
-    console.log(view.getInt16(1));
+    console.log(view.getInt32(1));
     
     if('world' in data)
     {
